feat: implement the GET /api/users endpoint
