+++ src/controller/userController.js
@@ -1,0 +1,6 @@
+const users = [];
+
+export const getUsers = (req, res) => {
+  res.writeHead(200, { 'Content-Type': 'application/json' });
+  res.end(JSON.stringify(users));
+};
